Factor out `ServerState` into a new file.
