+++ crates/sdk/src/lib.rs
@@ -4,6 +4,7 @@ pub mod default_main;
 pub mod fetch_metrics;
 pub mod json_rejection;
 pub mod json_response;
+pub mod state;
 pub mod tracing;
 
 pub use ndc_models as models;
